+++ smart-console/backend/src/main/java/com/smartdata/resto_console/model/Employee.java
@@ -458,21 +458,21 @@ public class Employee {
 
     // Convert JSON string to float[]
     @JsonIgnore
-    public float[] getFacialFeaturesAsArray() {
+    public double[] getFacialFeaturesAsArray() {
         try {
             if (!(facialFeatures == null || facialFeatures.trim().isEmpty())) {
-                return objectMapper.readValue(facialFeatures, float[].class);
+                return objectMapper.readValue(facialFeatures, double[].class);
             } else {
                 return null;
             }
         } catch (JsonProcessingException e) {
-            return new float[0]; // Handle conversion error
+            return new double[0]; // Handle conversion error
         }
     }
 
     // Convert float[] to JSON string before storing
     @JsonIgnore
-    public void setFacialFeaturesFromArray(float[] features) {
+    public void setFacialFeaturesFromArray(double[] features) {
         try {
             if (features != null) {
                 this.facialFeatures = objectMapper.writeValueAsString(features);
